Refactored model relationship slightly.
Took alot of the functionality only required for
one screen out of the central user model and
created a smaller one for that screen that
references the main model when required.
