+++ DMS/app/src/main/java/com/example/kbb12/dms/Model/InsulinModel.java
@@ -1,0 +1,14 @@
+package com.example.kbb12.dms.Model;
+
+import com.example.kbb12.dms.ErrorHandling.ErrorReadWriteModel;
+import com.example.kbb12.dms.Model.Insulin.DuplicateDoseException;
+import com.example.kbb12.dms.Model.Insulin.LongActingInsulinDose;
+
+import java.util.List;
+
+/**
+ * Created by kbb12 on 24/02/2017.
+ */
+public interface InsulinModel extends ErrorReadWriteModel {
+    public void saveDoses(List<LongActingInsulinDose> basicDoses,String longActingInsulinBrandName) throws DuplicateDoseException;
+}
